requirements.txt
anyio==3.6.1
APScheduler==3.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.0
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.1
orjson==3.8.0
psycopg2==2.9.4
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tweepy==4.10.1
typing_extensions==4.4.0
tzdata==2022.4
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
